Enable proper picking of the base images
[ci skip]
