+++ push.py
@@ -17,7 +17,7 @@ parser = argparse.ArgumentParser(description='Build local.')
 parser.add_argument('-b', '--branch', help="log choosen preCICE branch")
 parser.add_argument('-t', '--test', help="choose system tests you want to use")
 parser.add_argument('-s', '--success', action='store_true' ,help="only upload log file")
-parser.add_argument('-o', '--os', type=str,help="Variant used", default= "Ubuntu1604")
+parser.add_argument('--base', type=str,help="Base preCICE image used", default= "Ubuntu1604")
 
 args = parser.parse_args()
 
@@ -26,7 +26,7 @@ if __name__ == "__main__":
 
     # Creating new logfile. if it exists, truncate content.
     log = open("log_" + systest, "w")
-    foam_version = 4.1 if args.os == "1604" else 5
+    foam_version = 4.1
     # Saving versions of used software in system test systest.
     if systest == "of-ccx":
         log.write("OpenFOAM version: {}\n".format(foam_version))
@@ -50,7 +50,7 @@ if __name__ == "__main__":
         ccall(["echo preCICE Version: $(git ls-remote https://github.com/precice/precice.git | grep "+ args.branch +")"], stdout=log)
     else:
         ccall(["echo preCICE Version: $(git ls-remote --tags https://github.com/precice/precice.git | tail -1)"], stdout=log)
-    log.write("Distribution : {}\n".format(args.os))
+    log.write("Base preCICE image used : {}\n".format(args.base))
     # Saving current date in logfile.
     localtime = str(time.asctime(time.localtime(time.time())))
     log.write("System testing at " + localtime + "\n")
@@ -59,12 +59,12 @@ if __name__ == "__main__":
     # Pushing outputfiles and logfile to repo.
     # Clone repository.
     ccall("git clone https://github.com/precice/precice_st_output")
-    log_dir = os.getcwd() + "/precice_st_output/" + args.os
+    log_dir = os.getcwd() + "/precice_st_output/" + args.base
     ccall("mkdir -p " + log_dir)
     ccall("mv log_" + systest + " " + log_dir)
 
     if not args.success:
-        system_suffix = "." + args.os
+        system_suffix = "." + args.base
         # Move ouput to folder of this test case
         test_folder = os.getcwd() + '/Test_' + systest + system_suffix
         if not os.path.isdir(test_folder):
